Rename option.py to options.py
